Разработка панели администратора #8(добавлены чанки; мелкие правки)

+++ StoreCMS/Functions/ActionsWithDatabase/DeleteTemplate.cs
@@ -16,20 +16,22 @@ namespace Treynessen.Functions
         {
             if (!itemID.HasValue)
                 return;
-            List<UsualPage> usualPages = db.UsualPages.Where(up => up.TemplateId == itemID).ToList();
-            List<CategoryPage> categoryPages = db.CategoryPages.Where(cp => cp.TemplateId == itemID).ToList();
-            List<ProductPage> productPages = db.ProductPages.Where(pp => pp.TemplateId == itemID).ToList();
-            foreach (var up in usualPages)
+            Template template = db.Templates.FirstOrDefault(t => t.ID == itemID);
+            if (template == null)
+                return;
+            var usualPagesTask = db.UsualPages.Where(up => up.TemplateId == template.ID).ToListAsync();
+            var categoryPagesTask = db.CategoryPages.Where(cp => cp.TemplateId == template.ID).ToListAsync();
+            var productPagesTask = db.ProductPages.Where(pp => pp.TemplateId == template.ID).ToListAsync();
+            foreach (var up in usualPagesTask.Result)
                 up.Template = null;
-            db.UsualPages.UpdateRange(usualPages);
-            foreach (var cp in categoryPages)
+            db.UsualPages.UpdateRange(usualPagesTask.Result);
+            foreach (var cp in categoryPagesTask.Result)
                 cp.Template = null;
-            db.CategoryPages.UpdateRange(categoryPages);
-            foreach (var pp in productPages)
+            db.CategoryPages.UpdateRange(categoryPagesTask.Result);
+            foreach (var pp in productPagesTask.Result)
                 pp.Template = null;
-            db.ProductPages.UpdateRange(productPages);
-            Template template = db.Templates.FirstOrDefaultAsync(t => t.ID == itemID).Result;
             File.Delete($"{context.RequestServices.GetRequiredService<IHostingEnvironment>().ContentRootPath}/Views/Templates/{template.Name}.cshtml");
+            db.ProductPages.UpdateRange(productPagesTask.Result);
             db.Templates.Remove(template);
             db.SaveChanges();
         }
